requirements.txt
pyaccsharedmemory==1.0.0
pyserial==3.5
